requirements.txt
langchain
pymongo
pypdf
unstructured
pytesseract
pdf2image
pillow
ollama
tqdm
python-dotenv
streamlit